requirements.txt
# Automation & Web
playwright>=1.40.0

# Data Validation & Typing
pydantic>=2.5.0
pydantic-settings>=2.0.0

# AI & Orchestration
langchain>=0.1.0
langchain-google-genai>=0.0.1

# Utilities
python-dotenv>=1.0.0
pytest>=7.4.0
